add eslint-config
