+++ web/libs/eslint-config/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/eslint-config.ts';
