+++ web/nx.json
@@ -7,49 +7,55 @@
     "version": "22.1.3"
   },
   "namedInputs": {
-    "noMarkdown": [
-      "!{projectRoot}/**/*.md"
-    ]
+    "noMarkdown": ["!{projectRoot}/**/*.md"]
   },
   "tasksRunnerOptions": {
     "default": {
       "runner": "nx/tasks-runners/default",
       "options": {
-        "cacheableOperations": [
-          "build",
-          "test:unit",
-          "lint",
-          "knip"
-        ]
+        "cacheableOperations": ["build", "test:unit", "lint", "knip"]
       }
     }
   },
   "targetDefaults": {
     "build": {
-      "inputs": [
-        "noMarkdown",
-        "^noMarkdown"
-      ],
-      "dependsOn": [
-        "^build"
-      ]
+      "inputs": ["noMarkdown", "^noMarkdown"],
+      "dependsOn": ["^build"]
     },
     "test": {
-      "inputs": [
-        "noMarkdown",
-        "^noMarkdown"
-      ]
+      "inputs": ["noMarkdown", "^noMarkdown"]
     },
     "lint": {
-      "inputs": [
-        "noMarkdown",
-        "^noMarkdown"
-      ]
+      "inputs": ["noMarkdown", "^noMarkdown"]
     },
     "dev": {
-      "dependsOn": [
-        "^build"
-      ]
+      "dependsOn": ["^build"]
     }
-  }
+  },
+  "plugins": [
+    {
+      "plugin": "@nx/js/typescript",
+      "options": {
+        "typecheck": {
+          "targetName": "typecheck"
+        },
+        "build": {
+          "targetName": "build",
+          "configName": "tsconfig.lib.json",
+          "buildDepsName": "build-deps",
+          "watchDepsName": "watch-deps"
+        }
+      },
+      "exclude": ["libs/eslint-config/*"]
+    },
+    {
+      "plugin": "@nx/js/typescript",
+      "include": ["libs/eslint-config/*"],
+      "options": {
+        "typecheck": {
+          "targetName": "typecheck"
+        }
+      }
+    }
+  ]
 }
