+++ web/package.json
@@ -5,9 +5,18 @@
   "private": true,
   "packageManager": "pnpm@10.23.0",
   "devDependencies": {
-    "nx": "catalog:"
+    "@nx/js": "catalog:",
+    "@swc-node/register": "~1.9.1",
+    "@swc/core": "~1.5.7",
+    "@swc/helpers": "~0.5.11",
+    "@types/node": "20.19.9",
+    "nx": "catalog:",
+    "prettier": "^2.6.2",
+    "tslib": "^2.3.0",
+    "typescript": "~5.9.2"
   },
   "scripts": {
     "preinstall": "npx only-allow pnpm"
-  }
+  },
+  "dependencies": {}
 }
